+++ src/main/java/fuel/domain/K5.java
@@ -1,6 +1,6 @@
-package fuel;
+package fuel.domain;
 
-public class K5 extends Car{
+public class K5 extends Car {
 
     private static final String NAME = "K5";
     private static final int DISTANCE_PER_LITER = 13;
